refactor: 안드로이드 테스트 패키지 변경
