--- app/src/androidTest/java/woowacourse/movie/view/fixture/FakeContext.kt
@@ -1,6 +1,0 @@
-package woowacourse.movie.view.fixture
-
-import android.content.Context
-import androidx.test.core.app.ApplicationProvider
-
-val fakeContext: Context = ApplicationProvider.getApplicationContext()
